requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.4
sqlparse==0.4.3
django-cors-headers==3.13.0
six
apscheduler